+++ lib/redis_client/middlewares.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class RedisClient
+  module Middlewares
+    extend self
+
+    def call(command, _config)
+      yield command
+    end
+    alias_method :call_pipelined, :call
+  end
+end